+++ packages/preact/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix `valiField` and `valiForm` adapter for Valibot v0.13.0
+
 ## v0.6.1 (July 31, 2023)
 
 - Fix `valiField` and `valiForm` adapter for Valibot v0.7.0